Add files via upload
An Arduino Nano connected to an I2C OLED 128x32 screen shows the current voltage and an oscilloscope trace showing the history of the voltage.
This can easily be adapted to Arduino Uno without any change of code. The measurement is taken via a two-resistor 100kOhm voltage divider with one resistor between analog pin A7 and ground, and the other from the probe to A7. Any voltage on the probe will be divided by two so the analog pin will not rise above 5V with a 10V input.
Changing the resistor ratio will change the maximum voltage. Adding a zener 5V1 clamp will protect the analog pin in case of the probe touching a higher voltage.